requirements.txt

numpy

requests
python-dotenv
SQLAlchemy
matplotlib
pytest
mypy
streamlit
plotly
pandas
